requirements.txt
pyopengl==3.1.7
glfw==2.6.2
pyrr==0.10.3
pyglm==2.7.0
pywavefront==1.3.3